refactor: cleanup PSBTImport component
